requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
supabase>=1.0.3
tqdm>=4.66.1
